solar_backend/requirements.txt
django>=5.0
djangorestframework>=3.14
django-cors-headers>=4.0
djangorestframework-simplejwt>=5.3
numpy>=1.24
pandas>=2.0
scikit-learn>=1.3
tensorflow>=2.12.0
python-dotenv>=1.0
requests>=2.31
joblib>=1.3
h5py>=3.9
